pyproject.toml
[project]
name = "gdrive-tools"
version = "0.1.0"
description = "Foundational Google Drive and Docs functionality for automation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
